Get rid of redundant carousels
